src/requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
defusedxml==0.5.0
Django==2.2.9
django-braces==1.13.0
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.0.0
django-shotgun==0.2
djangorestframework==3.9.1
djangorestframework-csv==2.1.0
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.4.0
drf-amsterdam==0.1.8
drf-extensions==0.4.0
drf-hal-json==0.9.1
drf-nested-fields==0.9.4
drf-yasg==1.15.0
factory-boy==2.11.1
fake-factory==9999.9.9
Faker==1.0.1
flake8==3.6.0
graypy==0.3.1
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
packaging==18.0
pip-review==1.0
psycopg2-binary==2.7.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.7.1
pyparsing==2.3.1
pyslack-real==0.6.0
python-dateutil==2.7.5
pytz==2018.9
PyYAML==4.2b4
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sphinx-me==0.3
sqlparse==0.2.4
text-unidecode==1.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.2
Werkzeug==0.15.3
